requirements.txt
asgiref==3.8.1
astroid==3.3.5
black==24.10.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.0
pylint==3.3.1
python-dotenv==1.0.1
sqlparse==0.5.2
tomli==2.1.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
